Changed how default config and config seek work.
You can now run lab in two ways.  If you pass no arguments it will assume you are
working out of the current working directory.  Otherwise, you can pass a directory
as an argument as a starting point.  In both cases, lab will seek backwards as it always
has for a lab.config file.  If this file is not found by the time it hits root, it will use a
default built in config.